LPD-56133 Refactor requests to use ApiHelper from service folder
